chore: Added the login route logic for admin routes
This commit updates the .env.example file to include the `JWT_SECRET` variable. It also adds a new script `gen:key` to the `package.json` file, which generates a random 256-bit (32 bytes) key for JWT token signing. This key is used for securely signing and verifying JWT tokens in the application.