refactor: Update admin middleware to use findById instead of findOne
This commit updates the admin controller to sign only the user's ID in JWT token instead of signing the whole user object. It also updates the admin middleware to use the findById method instead of findOne for querying the database to check if an admin exists with the given ID. This change reduces the token length and improves security of tokens.